services/bot/requirements.txt
discord
asyncpraw
aiohttp
asyncio
helper
python-dotenv
gitpython
pycord
requests
pytz
ics
google-genai
pika
